+++ examples/nlp/text_normalization_as_tagging/dataset_preparation/prepare_corpora_after_alignment.py
@@ -24,8 +24,7 @@ from argparse import ArgumentParser
 from collections import Counter
 from typing import Dict, Optional, TextIO, Tuple
 
-from examples.nlp.text_normalization_as_tagging.dataset_preparation.utils import get_src_and_dst_for_alignment
-
+from nemo.collections.nlp.data.text_normalization_as_tagging.utils import get_src_and_dst_for_alignment
 from nemo.utils import logging
 
 parser = ArgumentParser(description="Produce data for the ThutmoseTaggerModel")
